backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
google-generativeai==0.7.1
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.0.0 